skip install on ubuntu if already installed

+++ src/index.js
@@ -9,7 +9,21 @@ const IS_MACOS = process.platform === 'darwin';
 const IS_LINUX_ARM = process.platform === 'linux' && process.arch === 'arm';
 const IS_LINUX_X64 = process.platform === 'linux' && process.arch === 'x64';
 
-function installAwsCliLinuxX64() {
+async function isAlreadyInstalled(toolName) {
+  const cachePath = await tc.find(toolName, '*');
+  const systemPath = await io.which(toolName);
+  if (cachePath) return cachePath;
+  if (systemPath) {
+    return systemPath;
+  }
+  return false;
+}
+
+async function installAwsCliLinuxX64() {
+  if (await isAlreadyInstalled("aws")) {
+    return resolve(`AWS CLI is already installed`);
+  }
+
   return new Promise((resolve, reject) => {
     exec('curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" && unzip awscliv2.zip && sudo ./aws/install', (error, stdout, stderr) => {
       if (error) {
@@ -23,7 +37,11 @@ function installAwsCliLinuxX64() {
   });
 }
 
-function installAwsCliLinuxARM() {
+async function installAwsCliLinuxARM() {
+  if (await isAlreadyInstalled("aws")) {
+    return resolve(`AWS CLI is already installed`);
+  }
+
   return new Promise((resolve, reject) => {
     exec('curl "https://awscli.amazonaws.com/awscli-exe-linux-aarch64.zip" -o "awscliv2.zip" && unzip awscliv2.zip && sudo ./aws/install', (error, stdout, stderr) => {
       if (error) {
@@ -40,7 +58,7 @@ function installAwsCliLinuxARM() {
 async function installAWSCliWindows() {
   const downloadUrl = 'https://s3.amazonaws.com/aws-cli/AWSCLISetup.exe';
   const tool = new toolManager(downloadUrl)
-  const isInstalled = await tool.isAlreadyInstalled('aws')
+  const isInstalled = await isAlreadyInstalled('aws')
   if (typeof isInstalled === 'string') {
     console.log('WARNING: AWS CLI is already installed but we shall continue');
   }
@@ -81,6 +99,6 @@ async function _installTool() {
   }
 }
 
-_installTool()
-
 module.exports = { _installTool };
+
+_installTool()
